added different states, logic for vending machine

+++ src/StateDrivenPattern/States/State.java
@@ -2,7 +2,10 @@ package StateDrivenPattern.States;
 
 public interface State {
 
-    public void insertCoin(int amount);
+    public void insertCoin(double amount);
+
+
+
     public void pressbutton(int aisleNumber);
 
     public void dispense(int aisleNumber);
